Fix
Moved Transforms for weapon positions around the player. Filled all weapons position in Player script. Changed weapon behaviour for aiming to the own nearest target and not the nearest target to the player. Now all enemies spawn as EnemySpawnerSystem's children. Fixed issue in which the spawner continues to spawn enemies due to inside timer management.